optional: show comment count on blog index
set disqus_show_comment_count to true to have disqus show the comment
count of each article on the blog index.

Additionally, if disqus_show_comment_count is not set, don't even
include the disqus script on the blog index as that one is only needed
for the comment counts.