Corrected error in new_post rake task when source/_posts doesn't exist
new_post now tests if the source posts dir exists and creates it if it
doesn't.

This will happen if, when reading the docs, the user runs the new_post
task on a fresh install, and we don't want new users to experience
errors on their first attempt to create a post.