Moved Twitter js into a namespace and removed ancillary functions from the global object.
Also removed jXHR code from here as it was being doubled up for the Github plugin. The Twitter & Github HTML partials now use:

  <script>window.jXHR || document.write('<script src="/javascripts/libs/jXHR.js"><\/script>')</script>

In order to ensure jXHR is present. Loading the external file asynchronously meant it wasn't present when the Twitter code was executing I found, hence the static loading with "pre-existing?" check.

A subjective one this but I personally prefer this style of date representation. I had to make a style sheet to change to accommodate the longer text.