Reads compass output path only from config.rb
Instead of hardwiring the css output path in compass_compiler.rb, it should
ONLY be taken from config.rb. For the generated file to be copied correctly by
jekyll, it must be placed in the source directory (fixed in config.rb).

In case you want to publish your CSS to a different directory, you only need
to change config.rb.