Add a Post-Scriptum to all posts.
This commit adds the option for putting a fixed text at the end of
each post, like a PS at the end of a letter.

Specifically, if you have configured the 'subscribe_email' then
a short text will invite the reader to subscribe to your mail list.

(This is useful, for example, to support inbound-marketing, where
you want to give a call-to-action to your readers, at the end of
each post.)

Additional custom text can be added in

  source/_includes/custom/post/postscriptum.html

Any such additional text will appear at the end of the generated
postscriptum.