Made it clearer that RVM vs rbenv was an either / or proposition
This is something that should be clear to everyone familiar with the
two, but if you aren't, and you are skimming the docs (as most developers
are wont to do) then it's easy to think you should install both.

This minor tweak to break out and bold the instructions that you can't have
both should take care of that.