In set_root_dir, update the site URL as well
Else a lot of absolute links are broken. This caused me some grief because
initially I wasn't sure whether the root would be automatically appended to
the site URL or not -- but since that's apparently not the case, we should
have the site URL be http://www.example.com/blog (if the root is "blog").