Pipes feed content through excerpt filter
When posts are enabled with the excerpt filter (using <!-- more -->)
the front page shows an abbreviated content but the feed shows the full
list.

This updates the main atom feed so that it shows the same abbreviated
content if it is available. A similar change is updated for the categories
as well.