Remove Gemfile.lock and add it to .gitignore
Gemfile.lock is personal to each developer, and it should not
be included in the repository.
This avoids conflicts with bundler when more than one version
of a gem is installed.
For example, currently the Gemfile.lock has rake version 0.9.2
locked but the Gemfile doesn't specify a version for rake.
When you run a fresh bundle install, it installs the currently
latest version 0.9.2.2 which will later cause an error message
when invoking rake without bundle exec, which should not be
necessary if runing inside an rvm gemset.
After this fix it's necessary to run bundle install to recreate
your personal Gemfile.lock according to the Gemfile specified
dependencies.

Signed-off-by: João Soares <jsoaresgeral@gmail.com>