Adds coffee-script support
Allows you to write JavaScript in coffeScript.

Make a MyScript.coffee and place YAML front matter in it:

    ---
    layout: nil
    ---
    alert "Hello CoffeeScript!"

and it will create a working MyScript.js

Because OctoPress post-filters content through RubyPants this script
adss a <script> wrapper around the content to prevent RubyPants
filtering.

See [this gist](https://gist.github.com/2925325) for the original
source.