Fixes URL sent to sharing services for both posts and pages.
Checks whether the destination is a post or a page and provides the correct Liquid tags to ensure the URL submitted to Twitter, Google + or Facebook is specific to that post or page (and not just the raw URL as previously).

i.e. a share button attached to a post on the blog index would send the root URL for the site instead of the URL specific to that post.