Make compiled scss debuggable when previewing
The Octopress configuration tells the scss to compile in a compressed
format, which is great for production. However, when developing a site
or a theme and previewing, it can be very helpful to be able to see the
compiled scss without all of the minification.

This commit sets, for the `preview` rake task, two options when calling
`compass compile` that should make working with scss a little easier for
developers. The first is `--debug-info` which includes metadata about
each selector that Chrome developer tools can use to give you more
information about the actual source of the style. The second is
`--output-style expanded` which prevents the minification step from
happening, allowing you to view the compiled CSS in a more readable
state.