Improve previewing
Previewing is typically done to verify a change to content, a plugin, or
a theme. In most of these cases, the person previewing will want to
re-generate before previewing. This commit adds a call to the `generate`
task from within the `preview` task to ensure that the newest content
will be previewed.

Additionally, since previewing is often done to verify changes to
themes, it can be helpful to be able to see un-minified assets. This
commit also sets a default value of 'development' to the `OCTOPRESS_ENV`
environment value, which is used when making decisions as to how assets
are compiled. This currently prevents the minification of the compiled
scss.

I also trimmed some trailing whitespace in the Rakefile.