Updating tests to reflect changes in semantics.
Previous fixups resulted in a semantic change wherein specifying an environment in config files didn't work anymore.

Unfortunately, it gets a bit sticky trying to restore that behavior robustly, and I'm not sure it's that valuable anyway.

Basically, the defaults for preview and generate are sane (development and production respectively), and OCTOPRESS_ENV is adhered to rigidly.