Begin expanding upon the environment functionality with per-env config files.
The currently extant environment functionality doesn't really afford one much in terms of capabilities -- this is a first step towards making it useful.

Key use-case for me is that I want the hostname to be different in development, staging, and production. (Workstation, Github Pages, and DreamHost or some other such system respectively...)