Update Rakefile to support hostname specification
While the typical case is to support `localhost`
on a laptop where DNS routing is not used,
assumption of this name is incompatible for those
with a remote-system / mainframe-esque work
environment.

This change surfaces to users that this is
changeable in case they have the ability to change
this and use name resolution (if they so desire).