allow scrolling for code blocks on iOS devices
Code blocks on iOS devices are not touch scrollable!
The only way to scroll code blocks on an iPhone is to try to touch the tiny
scroll bar and move it! In comparison to Android it is very complicated to view
octopress code!
To reproduce this error you need a native iOS device (or iOS simulator on OS X)
and open an octopress blog with Safari or Chrome.
A detailed description of this problem is described here:
http://drawingablank.me/blog/my-gripe-with-code-blocks.html